requirements.txt
aiofiles==24.1.0
aiomysql==0.2.0
aiosqlite==0.21.0
ansible-runner==2.4.0
bcrypt==3.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
html5tagger==1.3.0
httptools==0.6.4
idna==3.10
importlib-metadata==6.2.1
iso8601==2.1.0
jwt==1.3.1
lockfile==0.12.2
multidict==6.1.0
packaging==24.2
pexpect==4.9.0
ptyprocess==0.7.0
pycparser==2.22
PyMySQL==1.1.1
pypika-tortoise==0.5.0
python-daemon==3.1.2
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
sanic==24.12.0
sanic-routing==23.12.0
six==1.17.0
tortoise-orm==0.24.2
tracerite==1.1.1
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
uvloop==0.21.0
websockets==15.0.1
zipp==3.21.0
